[refactor] SellerOauthController의 로그인 및 토큰 관련 메서드를 OAuthController로 이동 - API 네이밍 변경 - 통합 및 단위 테스트 코드 수정 - Cookie 생성 메서드를 CookieUtils로 이동 - CookieFixture 추가
