+++ src/test/java/com/bbangle/bbangle/auth/seller/controller/SellerOauthControllerTest.java
@@ -5,13 +5,11 @@ import static org.hamcrest.Matchers.containsString;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
 import static org.mockito.Mockito.times;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.bbangle.bbangle.auth.oauth.client.dto.TokenResponse;
 import com.bbangle.bbangle.auth.seller.controller.dto.GenerateTokenRequest;
 import com.bbangle.bbangle.auth.seller.controller.dto.GenerateTokenResponse;
 import com.bbangle.bbangle.auth.seller.facade.OAuth2SellerFacade;
@@ -27,7 +25,6 @@ import com.bbangle.bbangle.config.security.jwt.TokenProvider;
 import com.bbangle.bbangle.exception.BbangleErrorCode;
 import com.bbangle.bbangle.exception.BbangleException;
 import com.bbangle.bbangle.seller.domain.model.CertificationStatus;
-import jakarta.servlet.http.Cookie;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -131,80 +128,5 @@ class SellerOauthControllerTest {
                 .content(jsonDataEncoder.encode(new GenerateTokenRequest(code)))
             )
             .andExpect(status().isUnauthorized());
-    }
-
-    @Test
-    @DisplayName("Refresh Token 쿠키가 존재하면 토큰을 재발급한다.")
-    void success_reissueToken() throws Exception {
-
-        // given
-        String refreshToken = "validRefreshToken";
-        Cookie cookie = new Cookie("refreshToken", refreshToken);
-
-        TokenResponse tokenResponse = TokenResponse.builder()
-            .refreshToken("newRefreshToken")
-            .accessToken("newAccessToken")
-            .build();
-
-        given(oAuth2SellerFacade.reissueToken(refreshToken)).willReturn(tokenResponse);
-
-        // when & then
-        mockMvc.perform(post(SellerApiPath.PREFIX + "/oauth2/reissue")
-            .cookie(cookie))
-            .andExpect(status().isOk())
-            .andExpect(header().string("Authorization", "Bearer newAccessToken"))
-            .andExpect(header().string(HttpHeaders.SET_COOKIE, containsString("refreshToken=newRefreshToken")))
-            .andExpect(jsonPath("$.success").value(true))
-            .andExpect(jsonPath("$.code").value(SUCCESS.getCode()))
-            .andExpect(jsonPath("$.message").value(SUCCESS.getMessage()));
-
-        then(responseService).should(times(1)).getSuccessResult();
-    }
-
-    @Test
-    @DisplayName("Refresh Token 쿠키가 없으면 401을 반환한다.")
-    void failure_reissueToken_notExistCookie() throws Exception {
-
-        // when & then
-        mockMvc.perform(post(SellerApiPath.PREFIX + "/oauth2/reissue"))
-            .andExpect(status().isUnauthorized());
-
-        // 쿠키가 없을 경우 oAuth2SellerFacade는 호출조차 하지 않음
-        then(oAuth2SellerFacade).shouldHaveNoInteractions();
-    }
-
-    @Test
-    @DisplayName("Refresh Token이 유효하지 않을 경우 401을 반환한다.")
-    void failure_reissueToken_invalidRefreshToken() throws Exception {
-
-        // given
-        String refreshToken = "invalidRefreshToken";
-        Cookie cookie = new Cookie("refreshToken", refreshToken);
-
-        given(oAuth2SellerFacade.reissueToken(refreshToken))
-            .willThrow(new BbangleException(BbangleErrorCode._UNAUTHORIZED));
-
-        // when & then
-        mockMvc.perform(post(SellerApiPath.PREFIX + "/oauth2/reissue")
-                .cookie(cookie))
-            .andExpect(status().isUnauthorized());
-    }
-
-    @Test
-    @DisplayName("로그아웃 시 항상 만료된 쿠키를 반환한다.")
-    void logout() throws Exception {
-
-        // given
-        Cookie cookie = new Cookie("refreshToken", "refreshToken");
-
-        // when & then
-        mockMvc.perform(delete(SellerApiPath.PREFIX + "/oauth2/logout")
-            .cookie(cookie))
-            .andExpect(status().isOk())
-            .andExpect(header().string(HttpHeaders.SET_COOKIE, containsString("refreshToken=")))
-            .andExpect(header().string(HttpHeaders.SET_COOKIE, containsString("Max-Age=0")))
-            .andExpect(jsonPath("$.success").value(true))
-            .andExpect(jsonPath("$.code").value(SUCCESS.getCode()))
-            .andExpect(jsonPath("$.message").value(SUCCESS.getMessage()));
     }
 }
